refactor: organize examples
