+++ editor/example/json-spec/basic-semantic-zoom.ts
