+++ editor/example/json-spec/cancer-variant.ts
