+++ editor/example/json-spec/circular-overview-linear-detail-views.ts
