+++ editor/example/json-spec/corces.ts
