+++ editor/example/json-spec/debug.ts
