+++ editor/example/json-spec/gene-annotation.ts
