+++ editor/example/json-spec/gene-transcript.ts
