+++ editor/example/json-spec/give.ts
