+++ editor/example/json-spec/gosling-data.ts
