+++ editor/example/json-spec/gremlin.ts
