+++ editor/example/json-spec/ideograms.ts
