+++ editor/example/json-spec/pathogenic.ts
