+++ editor/example/json-spec/responsive-alignment.ts
@@ -1,6 +1,6 @@
 import { OverlaidTracks, SingleTrack, SingleView } from '@gosling.schema';
 import type { GoslingSpec } from 'gosling.js';
-import { CHANNEL_DEFAULTS } from '../../src/core/channel';
+import { CHANNEL_DEFAULTS } from '../../../src/core/channel';
 
 type GapBarProp = {
     title: boolean;
