+++ editor/example/json-spec/responsive.ts
