+++ editor/example/json-spec/sars-cov-2.ts
