+++ editor/example/json-spec/sashimi.ts
