+++ editor/example/json-spec/semantic-zoom.ts
