+++ editor/example/json-spec/theme.ts
