+++ editor/example/json-spec/vertical-band.ts
