+++ editor/example/json-spec/visual-encoding.ts
